+++ graal/com.oracle.graal.truffle.hotspot/src/com/oracle/graal/truffle/hotspot/HotSpotOptimizedCallTarget.java
@@ -97,15 +97,13 @@ public final class HotSpotOptimizedCallTarget extends OptimizedCallTarget {
         if (m != null) {
             CompilerAsserts.neverPartOfCompilation();
             installedCode = null;
-            inliningResult = null;
             compilationProfile.reportInvalidated();
             logOptimizedInvalidated(this, oldNode, newNode, reason);
         }
         cancelInstalledTask(oldNode, newNode, reason);
     }
 
-    @Override
-    protected void cancelInstalledTask(Node oldNode, Node newNode, CharSequence reason) {
+    private void cancelInstalledTask(Node oldNode, Node newNode, CharSequence reason) {
         Future<InstalledCode> task = this.installedCodeTask;
         if (task != null) {
             task.cancel(true);
@@ -154,7 +152,7 @@ public final class HotSpotOptimizedCallTarget extends OptimizedCallTarget {
             return null;
         } else {
             performInlining();
-            cancelInlinedCallOptimization();
+// cancelInlinedCallOptimization();
             logOptimizingQueued(this);
             this.installedCodeTask = compiler.compile(this);
             if (!TruffleBackgroundCompilation.getValue()) {
@@ -181,8 +179,6 @@ public final class HotSpotOptimizedCallTarget extends OptimizedCallTarget {
                 }
             }
             return null;
-        } finally {
-            onCompilationDone();
         }
     }